Refactor: move sync logic into Editor_Sync_State, rename from Editor_Session_Guard [ED-24605]
